feat: Added 'Exit' option to user:edit and user:remove commands

+++ src/Command/User/RemoveUserCommand.php
@@ -69,6 +69,10 @@ class RemoveUserCommand extends Command
         while (!$userFound) {
             $user = $this->userHelper->getUser($io, 'Which user you want to remove?');
 
+            if ($user === null) {
+                break;
+            }
+
             $message = \sprintf(
                 'Is this the user [%s - %s (%s %s)] which you want to remove?',
                 $user->getId(),
@@ -81,14 +85,17 @@ class RemoveUserCommand extends Command
         }
 
         /** @var User $user */
+        if ($user instanceof User) {
+            // Delete user
+            $this->userResource->delete($user->getId());
 
-        // Delete user
-        $this->userResource->delete($user->getId());
+            $message = 'User removed - have a nice day';
+        } else {
+            $message = 'Nothing changed - have a nice day';
+        }
 
         if ($input->isInteractive()) {
-            $io->success([
-                'User removed - have a nice day',
-            ]);
+            $io->success($message);
         }
 
         return null;
